Removed all warnings i.e. -Werror compliant.
	* src/FootTrajectoryGeneration/LeftAndRightFootTrajectoryGenerationMultiple.cpp:
	int/uint comparison.
	* src/Mathematics/qld.cpp:
	unused variables.
	* src/PreviewControl/ZMPPreviewControlWithMultiBodyZMP.cpp:
	int/uint comparison.
	* src/ZMPRefTrajectoryGeneration/AnalyticalMorisawaCompact.cpp:
	variables used on debug mode.
	* src/ZMPRefTrajectoryGeneration/FilteringAnalyticalTrajectoryByPreviewControl.cpp:
	unused variables.
	* src/ZMPRefTrajectoryGeneration/ZMPConstrainedQPFastFormulation.cpp:
	int/uint comparison.
	* unitTesting/TestFootPrintPGInterface.cpp:
	unused variables.

+++ src/FootTrajectoryGeneration/LeftAndRightFootTrajectoryGenerationMultiple.cpp
@@ -553,7 +553,7 @@ InitializeFromRelativeSteps(deque<RelativeFootPosition> &RelativeFootPositions,
   if Continuity is set to true, and if the number of intervals so far is the number of
   intervals minus 2.
   */
-  if ((Continuity) && (IntervalIndex==m_DeltaTj.size()-2))
+  if ((Continuity) && (IntervalIndex==(int)(m_DeltaTj.size()-2)))
     {
       for(unsigned int lk=0;lk<2;lk++)
 	{
